labs/04-deploy-ai/01-backend-api/aais-lc-python-api/aais-lc-python/requirements.txt
azure-core==1.30.1
azure-identity==1.16.0
azure-search-documents==11.4.0
fastapi==0.110.1
uvicorn==0.25.0
openai==1.27.0
langchain==0.1.19
langchain-openai==0.1.3
tiktoken==0.6.0
python-dotenv==1.0.1
chainlit==1.0.506